Refactor: Create Traced transfer implementation
This class takes care of logging request and response headers as well as bodies
Note: This will be equivalent of using buffering for both request and response,
and will use more memory. Especially for production, do not just log to /dev/null,
instead, call setTrace(null) to ensure no logging is active!